@Lock(LockModeType.OPTIMISTIC_FORCE_INCREMENT)
interface Example<T> {
interface QueryByExampleExecutor<T> {

@Repository
class PassengerRepositoryImpl {

    @PersistenceContext //!!!!!!!!!!!
    private EntityManager entityManager;

    @Override
    public List<Passenger> findOrderedBySeatNumberLimitedTo(int limit) {
        return entityManager.createQuery("SELECT p FROM Passenger p ORDER BY p.seatNumber",
          Passenger.class).setMaxResults(limit) //!!!!!!!!!!!
          .getResultList();
    }
}

@EnableTransactionManagement

@Query(" FROM Event
         WHERE status = :status
             AND TIME_TO_SEC(TIMEDIFF(:now, lastUpdateTs)) >= :duration")
    List<PreviewEvent> findByStatusAndDuration(
        @Param("status") String status,
        @Param("duration") Integer duration,
        @Param("now") Timestamp now);

 SET
 SESSION // !!!!!!!!
 TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;

public boolean isDetached(Entity entity) {
    return entity.id != null  // must not be transient
        && !em.contains(entity)  // must not be managed now !!!!!!!!!!
        && em.find(Entity.class, entity.id) != null;  // must not have been removed
}

3.2 Version Attributes

The Version field or property is used by the persistence provider to perform optimistic locking. It is accessed and/or set by the persistence provider in the course of performing lifecycle operations on the entity instance. An entity is automatically enabled for optimistic locking if it has a property or field mapped with a Version mapping.

So if the entity is a versioned object, such as a @Version has been specified, then the default persistence provider will perform optimistic locking.

StaleObjectStateException:
Row was updated or deleted by another transaction (or unsaved-value mapping was incorrect)

SOLUTION:
entityManager.detach(entity);